+++ synthetic_data_generator/data_collect_modi.h
@@ -1,9 +1,9 @@
-//int data_collect_modi(int tid, int no_left_rcr, int left_offset, int no_right_rcr, int right_offset, int spacing, int row, int col, int thread, int *shot, float *gather, float *data, int flag0, int flag, int *rcr_position_collect, int n)
-int data_collect_modi(int tid, int ns, int grid_y, int thread, float *gather, float *data, int *rcr_position_collect, int n, int n_src, int n_rcr)
+//long long int data_collect_modi(long long int tid, long long int no_left_rcr, long long int left_offset, long long int no_right_rcr, long long int right_offset, long long int spacing, long long int row, long long int col, long long int thread, long long int *shot, float *gather, float *data, long long int flag0, long long int flag, long long int *rcr_position_collect, long long int n)
+long long int data_collect_modi(long long int tid, long long int ns, long long int grid_y, long long int thread, float *gather, float *data, long long int *rcr_position_collect, long long int n, long long int n_src, long long int n_rcr)
 {
-    for(int i = 0; i < ns; i++)
+    for(long long int i = 0; i < ns; i++)
     {
-        for(int j = 0; j < n_rcr; j++)
+        for(long long int j = 0; j < n_rcr; j++)
         {
             data[(tid * ns * n_rcr) + n_rcr * i + j] = gather[tid * ns * grid_y + rcr_position_collect[(n) * thread + (tid + n_src * j)] + grid_y * i];
         }
